requirements.txt
streamlit
groq
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
unstructured
